requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.39.13
botocore==1.39.13
Brotli==1.1.0
cbor==1.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
dataclasses-json==0.6.7
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
FixThaiPDF==0.2.1
FlagEmbedding==1.3.5
frozenlist==1.7.0
fsspec==2025.3.0
gradio==5.38.2
gradio_client==1.11.0
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.0
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ir_datasets==0.5.11
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.8
lxml==6.0.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.9.3
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
openai==1.97.1
orjson==3.11.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
peft==0.16.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pypdf==5.8.0
pythainlp==5.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
ruff==0.12.5
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
trec-car-tools==2.6
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
uvicorn==0.35.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9
zstandard==0.23.0
